fix: fixed search page providers
