+++ src/components/Search/Page.scss
