--- src/components/Search/useSearch.tsx
@@ -1,13 +1,0 @@
-import type {SearchConfig, SearchData} from './types';
-
-import {createContext, useContext} from 'react';
-
-export const SearchContext = createContext<SearchConfig | SearchData | null | undefined>(null);
-
-SearchContext.displayName = 'SearchContext';
-
-export const SearchProvider = SearchContext.Provider;
-
-export function useSearch() {
-    return useContext(SearchContext);
-}
