+++ src/components/index.ts
@@ -1,4 +1,4 @@
-export {SearchContext, SearchProvider, useSearch} from './Search/useSearch';
+export {SearchContext, SearchProvider, useSearch} from './Search';
 export {RouterContext, RouterProvider, useRouter} from './Router';
 
 export {App} from './App';
